+++ mmm_eval/metrics/__init__.py
@@ -7,6 +7,7 @@ from .accuracy_functions import (
     calculate_std_for_singular_values_across_cross_validation_folds,
     calculate_stds_for_series_across_cross_validation_folds,
 )
+from .metric_models import calculate_smape
 
 __all__ = [
     "calculate_means_for_series_across_cross_validation_folds",
@@ -14,4 +15,5 @@ __all__ = [
     "calculate_mean_for_singular_values_across_cross_validation_folds",
     "calculate_std_for_singular_values_across_cross_validation_folds",
     "calculate_absolute_percentage_change",
+    "calculate_smape",
 ]