+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] ; 2025-07-17
+
+### Added
+
+- SMAPE metric to accuracy and CV accuracy tests
+
+### Changed
+
+- All metrics with "percentage" in the name to be actual percentages, e.g. 15 instead of 0.15
+
+## [0.8.0] ; 2025-07-17
+
+### Changed
+
+- Changed holdout split strategy for accuracy test to fixed test set size
+
 ## [0.7.1] ; 2025-07-15
 
 ### Fixed